feat(fe): organize components under folder
