+++ frontend-client/app/_components/HeaderAuthPanel.tsx
